Updates fonts and UI elements for improved aesthetics
Updates the font family to Mochiy Pop One for headings and Fredoka for general text to enhance the visual appeal of the application.

Improves the Navbar category dropdown and updates the About page layout and content to enhance user experience and provide a clearer brand story.

Adds new fonts, restructures layout, and refactors text for a better user experience.